Add the boot simulator
The boot simulator is a wrapper around the bootloader that can be used
to test its ability to perform bad powerdown recovery.  The main program
is written in rust http://www.rust-lang.org/, and I recommend using
rustup https://www.rustup.rs/ to install the compiler.

Once installed:

    cargo build --release

should build the simulator using the source in the parent of this tree.
You can then run with ./target/release/bootsim to verify the recovery.