boot: Use proper write alignment for FRDM64-K64F
The K64F device always writes to flash 8 bytes at a time.  As such, it
is not possible to rewrite these bytes, so the status values have to be
spaced 8 bytes apart.  This increases the size of the status
significantly (to over 3k), but is needed to avoid "excessive wear" to
the flash device.